dao created for allentities and backend updated

+++ server/src/controllers/busController.js
@@ -1,13 +1,13 @@
-// const Bus = require("../models/BusDetails");
 const db = require("../models");
 const Bus = db.bus_details;
 const Apifeatures = require("../utils/apiFeatures");
-
 const createError = require("../utils/error");
 const { busSchema } = require("../utils/validationSchema");
+
 async function checkExists(bus_number) {
+  console.log(bus_number);
   const bus = await Bus.findAll({
-    where: { bus_number: bus_number },
+    where: { id: bus_number },
   });
   return bus.length > 0 ? true : false;
 }
@@ -16,18 +16,18 @@ const createBus = async (req, res, next) => {
   try {
     const result = await busSchema.validateAsync(req.body);
     console.log(result);
-    const status = await checkExists(result.bus_number);
+    const status = await checkExists(result.id);
     if (status)
       return next(
         createError(
           401,
-          `${result.bus_number} already exist please use another busNumber`
+          `${result.id} already exist please use another busNumber`
         )
       );
     const newBus = await Bus.create({
+      id: result.id,
       bus_name: result.bus_name,
       bus_type: result.bus_type,
-      bus_number: result.bus_number,
     });
     await newBus.save();
     res.status(200).send("Bus added successfully");
@@ -36,18 +36,31 @@ const createBus = async (req, res, next) => {
     next(err);
   }
 };
+
 const updateBus = async (req, res, next) => {
   try {
-    const updateBus = await Bus.update(
-      req.body,
-      { where: { id: req.params.id } },
-      { new: true, runValidator: true, useFindAndModify: false }
-    );
-    if (!updateBus) {
-      return next(createError(404, "Bus not found"));
+    // const updateBus = await Bus.update(
+    //   req.body,
+    //   { where: { id: req.params.id } },
+    //   { new: true, runValidator: true, useFindAndModify: false }
+    // );
+    // if (!updateBus) {
+    //   return next(createError(404, "Bus not found"));
+    // }
+    // const bus = await Bus.findOne({ where: { id: req.params.id } });
+    // res.status(200).json({ bus, success: true });
+    const busNumber = req.params.id;
+    const status = await checkExists(busNumber);
+    if (status) {
+      const data = await busSchema.validateAsync(req.body);
+      const bus = await Bus.update(data, { where: { id: busNumber } });
+      return res.json({
+        data: "Bus details updated successfully",
+        status: true,
+      });
+    } else {
+      return next(createError(422, "Error bus number does not exists"));
     }
-    const bus = await Bus.findOne({ where: { id: req.params.id } });
-    res.status(200).json({ bus, success: true });
   } catch (err) {
     next(err);
   }
@@ -55,31 +68,45 @@ const updateBus = async (req, res, next) => {
 
 const deleteBus = async (req, res, next) => {
   try {
-    await Bus.destroy({ where: { id: req.params.id } });
-    res.status(200).json("Bus has been deleted");
+    const status = await checkExists(req.params.id);
+    if (status) {
+      await Bus.destroy({ where: { id: req.params.id } });
+      return res.json({
+        data: "Bus details deleted successfully",
+        status: true,
+      });
+    } else {
+      return next(createError(422, "Bus number does not exists"));
+    }
   } catch (err) {
     next(err);
   }
 };
 
-const getBus = async (req, res, next) => {
-  try {
-    const bus = await Bus.findOne({ where: { id: req.params.id } });
-    res.status(200).json(bus);
-  } catch (err) {
-    next(err);
-  }
-};
+// const getBus = async (req, res, next) => {
+//   try {
+//     const bus = await Bus.findOne({ where: { id: req.params.id } });
+//     res.status(200).json(bus);
+//   } catch (err) {
+//     next(err);
+//   }
+// };
+
 const getBusByBusNumber = async (req, res, next) => {
   try {
-    const bus = await Bus.findOne({
-      where: { bus_number: req.params.bus_number },
-    });
-    res.status(200).json(bus);
+    const busNumber = req.params.id;
+    const status = await checkExists(busNumber);
+    if (status) {
+      const bus = await Bus.findAll({ where: { id: busNumber } });
+      return res.json({ data: bus, status: true });
+    } else {
+      return next(createError(422, "Bus number does not exits"));
+    }
   } catch (err) {
     next(err);
   }
 };
+
 const getBuses = async (req, res, next) => {
   try {
     const apiFeatures = new Apifeatures(Bus, req.query).filter();
@@ -95,6 +122,6 @@ module.exports = {
   getBusByBusNumber,
   updateBus,
   deleteBus,
-  getBus,
+  // getBus,
   getBuses,
 };
